changed update_trajectory to a particle method
